Refactor docs to conform to Diátaxis home page structure
We are required in our roadmap to update the home pages of our Store
documentations to adhere to the standard model as defined in the Diátaxis framework. This commit thus creates the "Tutorial", "How-to", "Explanation"
and "Reference" buckets, and redistribute existing content into them. No
existing content has been deleted or rewritten, only reshuffled.

Adding new content to the docs to fill in requirements defined in Diátaxis
that are currently missing is out of the scope of this commit. The "Explanation"
bucket is empty for now because salient explanation points are currently
subsumed into the step-by-step tutorials.

Solves: https://warthogs.atlassian.net/browse/SN-2192
What is Diátaxis: https://diataxis.fr