Use "--export-yaml" when backing up configuration.
Using the YAML format should be encouraged, since we support directly
importing it using the "--import-yaml" option. The previous option would
provide a backup of the values, but required multiple files and did not
have an easy import option. Newly added secrets in newer versions of the
snap would also have been excluded from the documentation, making it
hard to keep up-to-date.